refactor: Add price and oldPrice columns to Step entity
The code changes in `step.entity.ts` add two new columns, `price` and `oldPrice`, to the `Step` entity. These columns are of type `float` and will be used to store the price and old price values for a step. This enhancement allows for more detailed tracking and display of step prices in the application.